CR Load Image List nodes
