+++ animation_nodes/lists.py
@@ -274,6 +274,7 @@ class CR_TextListSimple:
         return (texts, show_help, )
  
 #---------------------------------------------------------------------------------------------------------------------#
+'''
 class CR_ImageList:
 
     @classmethod
@@ -334,8 +335,7 @@ class CR_ImageList:
         show_help = "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes/wiki/List-Nodes#cr-image-list"          
 
         return (images, show_help, )   
-
-
+'''
 #---------------------------------------------------------------------------------------------------------------------#
 class CR_ImageListSimple:
 
@@ -434,9 +434,8 @@ NODE_CLASS_MAPPINGS = {
     "CR Text List":CR_TextList,
     "CR Text List Simple":CR_TextListSimple,
     "CR Image List":CR_ImageList,    
-    "CR Image List Simple":CR_ImageListSimple,          
-    ### More
-    "CR Input Text List":CR_InputTextList,       
+    "CR Image List Simple":CR_ImageListSimple,
+    "CR Input Text List":CR_InputTextList,     
 }
 '''
