+++ config.py
@@ -13,7 +13,7 @@ color_mapping = {
     "brown": (160, 85, 15),
     "gray": (128, 128, 128),
     "lightgray": (211, 211, 211),
-    "darkgray": (169, 169, 169),
+    "darkgray": (102, 102, 102),
     "olive": (128, 128, 0),
     "lime": (0, 128, 0),
     "teal": (0, 128, 128),
