+++ dev_nodes/dev_nodes.py
@@ -6,6 +6,34 @@
 from ..categories import icons
 
 #---------------------------------------------------------------------------------------------------------------------#
+'''
+class CR_ListLength:
+
+    @classmethod
+    def INPUT_TYPES(s):
+    
+        input_dir = folder_paths.input_directory
+        image_folder = [name for name in os.listdir(input_dir) if os.path.isdir(os.path.join(input_dir,name))] 
+    
+        return {"required": {
+            "any_type": (any_type,),
+            }
+        }
+
+    RETURN_TYPES = ("INT", "STRING", )
+    RETURN_NAMES = ("INT", "show_help", )
+    OUTPUT_IS_LIST = (False, False)
+    FUNCTION = "get_length"
+    CATEGORY = icons.get("Comfyroll/List")
+
+    def get_length(self, any_type):
+
+        list_len = len(any_type)
+
+        show_help = "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes/wiki/Other-Nodes#cr-list-length"
+
+        return (list_len, show_help, )
+'''
 
 #---------------------------------------------------------------------------------------------------------------------#
 # MAPPINGS
