+++ nodes/nodes.py
@@ -12,326 +12,13 @@ import comfy.sd
 import json
 import folder_paths
 import typing as tg
-import random
 import datetime
 from PIL import Image
 from PIL.PngImagePlugin import PngInfo
 from pathlib import Path
-from .graphics_functions import random_hex_color, random_rgb
 from ..categories import icons
 
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.realpath(__file__)), "comfy"))
-
-#---------------------------------------------------------------------------------------------------------------------#
-# Aspect Ratio Nodes
-#---------------------------------------------------------------------------------------------------------------------#
-class CR_AspectRatioSD15:
-    def __init__(self):
-        pass
-
-    @classmethod
-    def INPUT_TYPES(s):
-    
-        aspect_ratios = ["custom",
-                         "1:1 square 512x512",
-                         "1:1 square 1024x1024",
-                         "2:3 portrait 512x768",
-                         "3:4 portrait 512x682",
-                         "3:2 landscape 768x512",
-                         "4:3 landscape 682x512",
-                         "16:9 cinema 910x512",
-                         "1.85:1 cinema 952x512",
-                         "2:1 cinema 1024x512",
-                         "2.39:1 anamorphic 1224x512"]
-               
-        return {
-            "required": {
-                "width": ("INT", {"default": 512, "min": 64, "max": 8192}),
-                "height": ("INT", {"default": 512, "min": 64, "max": 8192}),
-                "aspect_ratio": (aspect_ratios,),
-                "swap_dimensions": (["Off", "On"],),
-                "upscale_factor": ("FLOAT", {"default": 1.0, "min": 0.1, "max": 100.0, "step":0.1}),
-                "batch_size": ("INT", {"default": 1, "min": 1, "max": 64})
-            }
-        }
-    RETURN_TYPES = ("INT", "INT", "FLOAT", "INT", "LATENT", "STRING", )
-    RETURN_NAMES = ("width", "height", "upscale_factor", "batch_size", "empty_latent", "show_help", )
-    FUNCTION = "Aspect_Ratio"
-    CATEGORY = icons.get("Comfyroll/Aspect Ratio")
-
-    def Aspect_Ratio(self, width, height, aspect_ratio, swap_dimensions, upscale_factor, batch_size):
-        if aspect_ratio == "2:3 portrait 512x768":
-            width, height = 512, 768
-        elif aspect_ratio == "3:2 landscape 768x512":
-            width, height = 768, 512
-        elif aspect_ratio == "1:1 square 512x512":
-            width, height = 512, 512
-        elif aspect_ratio == "1:1 square 1024x1024":
-            width, height = 1024, 1024
-        elif aspect_ratio == "16:9 cinema 910x512":
-            width, height = 910, 512
-        elif aspect_ratio == "3:4 portrait 512x682":
-            width, height = 512, 682
-        elif aspect_ratio == "4:3 landscape 682x512":
-            width, height = 682, 512
-        elif aspect_ratio == "1.85:1 cinema 952x512":            
-            width, height = 952, 512
-        elif aspect_ratio == "2:1 cinema 1024x512":
-            width, height = 1024, 512
-        elif aspect_ratio == "2.39:1 anamorphic 1224x512":
-            width, height = 1224, 512
-
-        if swap_dimensions == "On":
-            width, height = height, width
-           
-        latent = torch.zeros([batch_size, 4, height // 8, width // 8])
-
-        show_help = "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes/wiki/Aspect-Ratio-Nodes#cr-sd15-aspect-ratio"
-           
-        return(width, height, upscale_factor, batch_size, {"samples":latent}, show_help, )   
-
-#---------------------------------------------------------------------------------------------------------------------#
-class CR_SDXLAspectRatio:
-    def __init__(self):
-        pass
-
-    @classmethod
-    def INPUT_TYPES(s):
-    
-        aspect_ratios = ["custom",
-                                  "1:1 square 1024x1024",
-                                  "3:4 portrait 896x1152",
-                                  "5:8 portrait 832x1216",
-                                  "9:16 portrait 768x1344",
-                                  "9:21 portrait 640x1536",
-                                  "4:3 landscape 1152x896",
-                                  "3:2 landscape 1216x832",
-                                  "16:9 landscape 1344x768",
-                                  "21:9 landscape 1536x640"]
-        
-        return {
-            "required": {
-                "width": ("INT", {"default": 1024, "min": 64, "max": 8192}),
-                "height": ("INT", {"default": 1024, "min": 64, "max": 8192}),
-                "aspect_ratio": (aspect_ratios,),
-                "swap_dimensions": (["Off", "On"],),
-                "upscale_factor": ("FLOAT", {"default": 1.0, "min": 0.1, "max": 100.0, "step":0.1}),
-                "batch_size": ("INT", {"default": 1, "min": 1, "max": 64})
-            }
-        }
-    RETURN_TYPES = ("INT", "INT", "FLOAT", "INT", "LATENT", "STRING", )
-    RETURN_NAMES = ("width", "height", "upscale_factor", "batch_size", "empty_latent", "show_help", )
-    FUNCTION = "Aspect_Ratio"
-    CATEGORY = icons.get("Comfyroll/Aspect Ratio")
-
-    def Aspect_Ratio(self, width, height, aspect_ratio, swap_dimensions, upscale_factor, batch_size):
-        if aspect_ratio == "1:1 square 1024x1024":
-            width, height = 1024, 1024
-        elif aspect_ratio == "3:4 portrait 896x1152":
-            width, height = 896, 1152
-        elif aspect_ratio == "5:8 portrait 832x1216":
-            width, height = 832, 1216
-        elif aspect_ratio == "9:16 portrait 768x1344":
-            width, height = 768, 1344
-        elif aspect_ratio == "9:21 portrait 640x1536":
-            width, height = 640, 1536
-        elif aspect_ratio == "4:3 landscape 1152x896":
-            width, height = 1152, 896
-        elif aspect_ratio == "3:2 landscape 1216x832":
-            width, height = 1216, 832
-        elif aspect_ratio == "16:9 landscape 1344x768":
-            width, height = 1344, 768
-        elif aspect_ratio == "21:9 landscape 1536x640":
-            width, height = 1536, 640
-
-        if swap_dimensions == "On":
-            width, height = height, width
-             
-        latent = torch.zeros([batch_size, 4, height // 8, width // 8])
-
-        show_help = "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes/wiki/Aspect-Ratio-Nodes#cr-sdxl-aspect-ratio"
-           
-        return(width, height, upscale_factor, batch_size, {"samples":latent}, show_help, )  
-
-#---------------------------------------------------------------------------------------------------------------------#
-class CR_AspectRatio:
-    def __init__(self):
-        pass
-
-    @classmethod
-    def INPUT_TYPES(s):
-    
-        aspect_ratios = ["custom",
-                         "SD1.5 - 1:1 square 512x512",
-                         "SD1.5 - 2:3 portrait 512x768",
-                         "SD1.5 - 3:4 portrait 512x682",
-                         "SD1.5 - 3:2 landscape 768x512",
-                         "SD1.5 - 4:3 landscape 682x512",
-                         "SD1.5 - 16:9 cinema 910x512",
-                         "SD1.5 - 1.85:1 cinema 952x512",
-                         "SD1.5 - 2:1 cinema 1024x512",
-                         "SDXL - 1:1 square 1024x1024",
-                         "SDXL - 3:4 portrait 896x1152",
-                         "SDXL - 5:8 portrait 832x1216",
-                         "SDXL - 9:16 portrait 768x1344",
-                         "SDXL - 9:21 portrait 640x1536",
-                         "SDXL - 4:3 landscape 1152x896",
-                         "SDXL - 3:2 landscape 1216x832",
-                         "SDXL - 16:9 landscape 1344x768",
-                         "SDXL - 21:9 landscape 1536x640"]
-               
-        return {
-            "required": {
-                "width": ("INT", {"default": 1024, "min": 64, "max": 8192}),
-                "height": ("INT", {"default": 1024, "min": 64, "max": 8192}),
-                "aspect_ratio": (aspect_ratios,),
-                "swap_dimensions": (["Off", "On"],),
-                "upscale_factor": ("FLOAT", {"default": 1.0, "min": 0.1, "max": 100.0, "step":0.1}),
-                "prescale_factor": ("FLOAT", {"default": 1.0, "min": 0.1, "max": 100.0, "step":0.1}),
-                "batch_size": ("INT", {"default": 1, "min": 1, "max": 64})
-            }
-        }
-    RETURN_TYPES = ("INT", "INT", "FLOAT", "FLOAT", "INT", "LATENT", "STRING", )
-    RETURN_NAMES = ("width", "height", "upscale_factor", "prescale_factor", "batch_size", "empty_latent", "show_help", )
-    FUNCTION = "Aspect_Ratio"
-    CATEGORY = icons.get("Comfyroll/Aspect Ratio")
-
-    def Aspect_Ratio(self, width, height, aspect_ratio, swap_dimensions, upscale_factor, prescale_factor, batch_size):
-        
-        # SD1.5
-        if aspect_ratio == "SD1.5 - 1:1 square 512x512":
-            width, height = 512, 512
-        elif aspect_ratio == "SD1.5 - 2:3 portrait 512x768":
-            width, height = 512, 768
-        elif aspect_ratio == "SD1.5 - 16:9 cinema 910x512":
-            width, height = 910, 512
-        elif aspect_ratio == "SD1.5 - 3:4 portrait 512x682":
-            width, height = 512, 682
-        elif aspect_ratio == "SD1.5 - 3:2 landscape 768x512":
-            width, height = 768, 512    
-        elif aspect_ratio == "SD1.5 - 4:3 landscape 682x512":
-            width, height = 682, 512
-        elif aspect_ratio == "SD1.5 - 1.85:1 cinema 952x512":            
-            width, height = 952, 512
-        elif aspect_ratio == "SD1.5 - 2:1 cinema 1024x512":
-            width, height = 1024, 512
-        elif aspect_ratio == "SD1.5 - 2.39:1 anamorphic 1224x512":
-            width, height = 1224, 512 
-        # SDXL   
-        if aspect_ratio == "SDXL - 1:1 square 1024x1024":
-            width, height = 1024, 1024
-        elif aspect_ratio == "SDXL - 3:4 portrait 896x1152":
-            width, height = 896, 1152
-        elif aspect_ratio == "SDXL - 5:8 portrait 832x1216":
-            width, height = 832, 1216
-        elif aspect_ratio == "SDXL - 9:16 portrait 768x1344":
-            width, height = 768, 1344
-        elif aspect_ratio == "SDXL - 9:21 portrait 640x1536":
-            width, height = 640, 1536
-        elif aspect_ratio == "SDXL - 4:3 landscape 1152x896":
-            width, height = 1152, 896
-        elif aspect_ratio == "SDXL - 3:2 landscape 1216x832":
-            width, height = 1216, 832
-        elif aspect_ratio == "SDXL - 16:9 landscape 1344x768":
-            width, height = 1344, 768
-        elif aspect_ratio == "SDXL - 21:9 landscape 1536x640":
-            width, height = 1536, 640                
-        
-        if swap_dimensions == "On":
-            width, height = height, width
-        
-        width = int(width*prescale_factor)
-        height = int(height*prescale_factor)
-        
-        latent = torch.zeros([batch_size, 4, height // 8, width // 8])
-
-        show_help = "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes/wiki/Aspect-Ratio-Nodes#cr-aspect-ratio"
-           
-        return(width, height, upscale_factor, prescale_factor, batch_size, {"samples":latent}, show_help, )    
-
-#---------------------------------------------------------------------------------------------------------------------#
-class CR_AspectRatioBanners:
-    def __init__(self):
-        pass
-
-    @classmethod
-    def INPUT_TYPES(s):
-    
-        aspect_ratios = ["custom",
-                         "Large Rectangle - 336x280", 
-                         "Medium Rectangle - 300x250", 
-                         "Small Rectangle - 180x150",
-                         "Square - 250x250", 
-                         "Small Square - 200x200",
-                         "Button - 125x125", 
-                         "Half Page - 300x600",
-                         "Vertical Banner - 120x240", 
-                         "Wide Skyscraper - 160x600", 
-                         "Skyscraper - 120x600", 
-                         "Billboard - 970x250", 
-                         "Portrait - 300x1050", 
-                         "Banner - 468x60", 
-                         "Leaderboard - 728x90"]
-                                 
-        return {
-            "required": {
-                "width": ("INT", {"default": 1024, "min": 64, "max": 8192}),
-                "height": ("INT", {"default": 1024, "min": 64, "max": 8192}),
-                "aspect_ratio": (aspect_ratios,),
-                "swap_dimensions": (["Off", "On"],),
-                "upscale_factor": ("FLOAT", {"default": 1.0, "min": 0.1, "max": 100.0, "step":0.1}),
-                "prescale_factor": ("FLOAT", {"default": 1.0, "min": 0.1, "max": 100.0, "step":0.1}),
-                "batch_size": ("INT", {"default": 1, "min": 1, "max": 64})
-            }
-        }
-    RETURN_TYPES = ("INT", "INT", "FLOAT", "FLOAT", "INT", "LATENT", "STRING", )
-    RETURN_NAMES = ("width", "height", "upscale_factor", "prescale_factor", "batch_size", "empty_latent", "show_help", )
-    FUNCTION = "Aspect_Ratio"
-    CATEGORY = icons.get("Comfyroll/Aspect Ratio")
-
-    def Aspect_Ratio(self, width, height, aspect_ratio, swap_dimensions, upscale_factor, prescale_factor, batch_size):
-        
-        # Banner sizes
-        if aspect_ratio == "Large Rectangle - 336x280":
-            width, height = 336, 280
-        elif aspect_ratio == "Medium Rectangle - 300x250":
-            width, height = 300, 250
-        elif aspect_ratio == "Small Rectangle - 180x150":
-            width, height = 180, 150
-        elif aspect_ratio == "Square - 250x250":
-            width, height = 250, 250
-        elif aspect_ratio == "Small Square - 200x200":
-            width, height = 200	, 200
-        elif aspect_ratio == "Button - 125x125":
-            width, height = 125	, 125
-        elif aspect_ratio == "Half Page - 300x600":
-            width, height = 300, 600
-        elif aspect_ratio == "Vertical Banner - 120x240":
-            width, height = 120, 240
-        elif aspect_ratio == "Wide Skyscraper - 160x600":
-            width, height = 160, 600
-        elif aspect_ratio == "Skyscraper - 120x600":
-            width, height = 120, 600
-        elif aspect_ratio == "Billboard - 970x250":
-            width, height = 970, 250
-        elif aspect_ratio == "Portrait - 300x1050":
-            width, height = 300, 1050
-        elif aspect_ratio == "Banner - 468x60":
-            width, height = 168, 60
-        elif aspect_ratio == "Leaderboard - 728x90":
-            width, height = 728, 90              
-        
-        if swap_dimensions == "On":
-            width, height = height, width
-        
-        width = int(width*prescale_factor)
-        height = int(height*prescale_factor)
-        
-        latent = torch.zeros([batch_size, 4, height // 8, width // 8])
-
-        show_help = "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes/wiki/Aspect-Ratio-Nodes#cr-aspect-ratio-banners"
-           
-        return(width, height, upscale_factor, prescale_factor, batch_size, {"samples":latent}, show_help, ) 
 
 #---------------------------------------------------------------------------------------------------------------------#
 # Other Nodes
@@ -709,120 +396,13 @@ class CR_SelectModel:
         show_help = "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes/wiki/Other-Nodes#cr-select-model"
             
         return (model, clip, vae, model_name, show_help, )
-
-#---------------------------------------------------------------------------------------------------------------------#
-# List Nodes                
-#---------------------------------------------------------------------------------------------------------------------#
-class AnyType(str):
-    """A special type that can be connected to any other types. Credit to pythongosssss"""
-
-    def __ne__(self, __value: object) -> bool:
-        return False
-
-any_type = AnyType("*")
-
-class CR_FontFileList:
-
-    @classmethod
-    def INPUT_TYPES(s):
-    
-        comfyroll_font_dir = os.path.join(os.path.dirname(os.path.dirname(os.path.realpath(__file__))), "fonts")       
-        comfyroll_file_list = [f for f in os.listdir(comfyroll_font_dir) if os.path.isfile(os.path.join(comfyroll_font_dir, f)) and f.lower().endswith(".ttf")]
-
-        sources = ["System", "Comfyroll"]
-        
-        return {"required": {"source_folder": (sources,),
-                             #"full_folder_path": 
-                             "start_index": ("INT", {"default": 0, "min": 0, "max": 9999}),
-                             "max_rows": ("INT", {"default": 1000, "min": 1, "max": 9999}),                            
-                            }
-        }
-
-    RETURN_TYPES = (any_type, "STRING", )
-    RETURN_NAMES = ("LIST", "show_help", )
-    OUTPUT_IS_LIST = (True,)
-    FUNCTION = "make_list"
-    CATEGORY = icons.get("Comfyroll/List")
-
-    def make_list(self, source_folder, start_index, max_rows):
-
-        if source_folder == "System":
-            system_root = os.environ.get('SystemRoot')
-            system_font_dir = os.path.join(system_root, 'Fonts')   
-            file_list = [f for f in os.listdir(system_font_dir) if os.path.isfile(os.path.join(system_font_dir, f)) and f.lower().endswith(".ttf")]
-            #print(len(file_list))
-        elif source_folder == "Comfyroll":
-            comfyroll_font_dir = os.path.join(os.path.dirname(os.path.dirname(os.path.realpath(__file__))), "fonts")       
-            file_list = [f for f in os.listdir(comfyroll_font_dir) if os.path.isfile(os.path.join(comfyroll_font_dir, f)) and f.lower().endswith(".ttf")]
-            #print(len(comfyroll_file_list))
-        else:
-            pass
-        
-        # Ensure start_index is within the bounds of the list
-        start_index = max(0, min(start_index, len(file_list) - 1))
-
-        # Calculate the end index based on max_rows
-        end_index = min(start_index + max_rows, len(file_list))
-
-        # Extract the desired portion of the list
-        selected_files = file_list[start_index:end_index]
-
-        show_help = "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes/wiki/Other-Nodes#cr-font-file-list"
-
-        return (selected_files, show_help, )
-
-#---------------------------------------------------------------------------------------------------------------------#
-class CR_TextList:
-
-    @classmethod
-    def INPUT_TYPES(s):
-    
-        return {"required": {"multiline_text": ("STRING", {"multiline": False, "default": "text"}),
-                             "start_index": ("INT", {"default": 0, "min": 0, "max": 9999}),
-                             "max_rows": ("INT", {"default": 1000, "min": 1, "max": 9999}),
-                            }
-        }
-
-    RETURN_TYPES = ("STRING", "STRING", )
-    RETURN_NAMES = ("STRING", "show_help", )
-    OUTPUT_IS_LIST = (True,)
-    FUNCTION = "make_list"
-    CATEGORY = icons.get("Comfyroll/List")
-
-    def make_list(self, multiline_text, start_index, max_rows):
-
-        lines = multiline_text.split('\n')
-
-        # Ensure start_index is within the bounds of the list
-        start_index = max(0, min(start_index, len(lines) - 1))
-
-        # Calculate the end index based on max_rows
-        end_index = min(start_index + max_rows, len(lines))
-
-        # Extract the desired portion of the list
-        selected_rows = lines[start_index:end_index]
-
-        # Join the selected portion into a multiline string
-        text_list = "\n".join(selected_rows)
-
-        show_help = "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes/wiki/Other-Nodes#cr-text-list"
-
-        return (selected_rows, show_help, )
-
+       
 #---------------------------------------------------------------------------------------------------------------------#
 # MAPPINGS
 #---------------------------------------------------------------------------------------------------------------------#
 # For reference only, actual mappings are in __init__.py
 '''
 NODE_CLASS_MAPPINGS = {
-    ### Aspect ratio
-    "CR SD1.5 Aspect Ratio": CR_AspectRatioSD15,
-    "CR SDXL Aspect Ratio": CR_SDXLAspectRatio,
-    "CR Aspect Ratio": CR_AspectRatio,
-    "CR Aspect Ratio Banners":CR_AspectRatioBanners,
-    ### List nodes
-    "CR Font File List": CR_FontFileList,  
-    "CR Text List": CR_TextList, 
     ### Other
     "CR Image Output": CR_ImageOutput,
     "CR Integer Multiple": CR_IntegerMultipleOf,
@@ -832,7 +412,7 @@ NODE_CLASS_MAPPINGS = {
     "CR Split String": CR_SplitString,
     "CR Value": CR_Value,
     "CR Conditioning Mixer": CR_ConditioningMixer,
-    "CR Select Model": CR_SelectModel,
+    "CR Select Model": CR_SelectModel,    
 }
 '''
